model-services/base-diffusion/requirements.txt
fastapi
uvicorn
google-generativeai
pymongo
python-multipart
Pillow
requests
